Preserve Wispr timeout transaction identity
